tests/krate/publish: Extract `auth` module
