+++ src/tests/krate/publish/mod.rs
@@ -1,18 +1,17 @@
 use crate::builders::{CrateBuilder, DependencyBuilder, PublishBuilder};
 use crate::util::{RequestHelper, TestApp};
-use crates_io::controllers::krate::publish::{
-    missing_metadata_error_message, MISSING_RIGHTS_ERROR_MESSAGE,
-};
+use crates_io::controllers::krate::publish::missing_metadata_error_message;
 use crates_io::models::krate::MAX_NAME_LENGTH;
-use crates_io::schema::{api_tokens, versions_published_by};
+use crates_io::schema::versions_published_by;
 use crates_io::views::GoodCrate;
 use crates_io_tarball::TarballBuilder;
-use diesel::{ExpressionMethods, QueryDsl, RunQueryDsl};
+use diesel::{QueryDsl, RunQueryDsl};
 use http::StatusCode;
 use std::collections::BTreeMap;
 use std::iter::FromIterator;
 
 mod audit_action;
+mod auth;
 mod build_metadata;
 mod categories;
 mod dependencies;
@@ -55,38 +54,6 @@ fn uploading_new_version_touches_crate() {
     let updated_at_after = json.krate.updated_at;
 
     assert_ne!(updated_at_before, updated_at_after);
-}
-
-#[test]
-fn new_wrong_token() {
-    let (app, anon, _, token) = TestApp::full().with_token();
-
-    // Try to publish without a token
-    let crate_to_publish = PublishBuilder::new("foo", "1.0.0");
-    let response = anon.publish_crate(crate_to_publish);
-    assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    assert_eq!(
-        response.into_json(),
-        json!({ "errors": [{ "detail": "must be logged in to perform that action" }] })
-    );
-
-    // Try to publish with the wrong token (by changing the token in the database)
-    app.db(|conn| {
-        diesel::update(api_tokens::table)
-            .set(api_tokens::token.eq(b"bad" as &[u8]))
-            .execute(conn)
-            .unwrap();
-    });
-
-    let crate_to_publish = PublishBuilder::new("foo", "1.0.0");
-    let response = token.publish_crate(crate_to_publish);
-    assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    assert_eq!(
-        response.into_json(),
-        json!({ "errors": [{ "detail": "must be logged in to perform that action" }] })
-    );
-
-    assert!(app.stored_files().is_empty());
 }
 
 #[test]
@@ -215,29 +182,6 @@ fn new_krate_twice() {
         "index/fo/o_/foo_twice",
     ];
     assert_eq!(app.stored_files(), expected_files);
-}
-
-#[test]
-fn new_krate_wrong_user() {
-    let (app, _, user) = TestApp::full().with_user();
-
-    app.db(|conn| {
-        // Create the foo_wrong crate with one user
-        CrateBuilder::new("foo_wrong", user.as_model().id).expect_build(conn);
-    });
-
-    // Then try to publish with a different user
-    let another_user = app.db_new_user("another").db_new_token("bar");
-    let crate_to_publish = PublishBuilder::new("foo_wrong", "2.0.0");
-
-    let response = another_user.publish_crate(crate_to_publish);
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.into_json(),
-        json!({ "errors": [{ "detail": MISSING_RIGHTS_ERROR_MESSAGE }] })
-    );
-
-    assert!(app.stored_files().is_empty());
 }
 
 #[test]
